pyproject.toml
[project]
name = "document-qa-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.136.1",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-groq>=0.3.2",
    "langchain-huggingface>=0.1.2",
    "sentence-transformers>=3.0.0",
    "pypdf>=6.10.2",
    "python-docx>=1.2.0",
    "tiktoken>=0.12.0",
    "uvicorn>=0.46.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.20",
    "docx2txt>=0.8",
]
